+++ RDKShell/RDKShell.h
@@ -26,7 +26,9 @@
 #include <rdkshell/linuxkeys.h>
 #include <interfaces/ICapture.h>
 #include "tptimer.h"
-
+#ifdef ENABLE_RIALTO_FEATURE
+#include "RialtoConnector.h"
+#endif
 namespace WPEFramework {
 
     namespace Plugin {
@@ -142,6 +144,10 @@ namespace WPEFramework {
             static const string RDKSHELL_METHOD_KEY_REPEAT_CONFIG;
             static const string RDKSHELL_METHOD_GET_GRAPHICS_FRAME_RATE;
             static const string RDKSHELL_METHOD_SET_GRAPHICS_FRAME_RATE;
+#ifdef HIBERNATE_SUPPORT_ENABLED
+            static const string RDKSHELL_METHOD_CHECKPOINT;
+            static const string RDKSHELL_METHOD_RESTORE;
+#endif
 
             // events
             static const string RDKSHELL_EVENT_ON_USER_INACTIVITY;
@@ -164,6 +170,10 @@ namespace WPEFramework {
             static const string RDKSHELL_EVENT_ON_EASTER_EGG;
             static const string RDKSHELL_EVENT_ON_WILL_DESTROY;
             static const string RDKSHELL_EVENT_ON_SCREENSHOT_COMPLETE;
+#ifdef HIBERNATE_SUPPORT_ENABLED
+            static const string RDKSHELL_EVENT_ON_CHECKPOINTED;
+            static const string RDKSHELL_EVENT_ON_RESTORED;
+#endif
 
             void notify(const std::string& event, const JsonObject& parameters);
             void pluginEventHandler(const JsonObject& parameters);
@@ -258,6 +268,10 @@ namespace WPEFramework {
             uint32_t keyRepeatConfigWrapper(const JsonObject& parameters, JsonObject& response);
             uint32_t getGraphicsFrameRateWrapper(const JsonObject& parameters, JsonObject& response);
             uint32_t setGraphicsFrameRateWrapper(const JsonObject& parameters, JsonObject& response);
+#ifdef HIBERNATE_SUPPORT_ENABLED
+            uint32_t checkpointWrapper(const JsonObject& parameters, JsonObject& response);
+            uint32_t restoreWrapper(const JsonObject& parameters, JsonObject& response);
+#endif
 
         private/*internal methods*/:
             RDKShell(const RDKShell&) = delete;
@@ -399,7 +413,19 @@ namespace WPEFramework {
 
               private:
                   virtual void StateChange(PluginHost::IShell* shell);
+#ifdef USE_THUNDER_R4
+                  void Activation(const string& name, PluginHost::IShell* plugin) override
+                  {
+                  }
 
+                  void Deactivation(const string& name, PluginHost::IShell* plugin) override
+                  {
+                  }
+
+                  virtual void  Activated(const string& callSign,  PluginHost::IShell* plugin);
+                  virtual void  Deactivated(const string& callSign,  PluginHost::IShell* plugin);
+                  virtual void  Unavailable(const string& callSign,  PluginHost::IShell* plugin);
+#endif /* USE_THUNDER_R4 */
               private:
                   RDKShell& mShell;
             };
@@ -440,6 +466,9 @@ namespace WPEFramework {
             bool mEnableEasterEggs;
             ScreenCapture mScreenCapture;
             bool mErmEnabled;
+#ifdef ENABLE_RIALTO_FEATURE
+        std::shared_ptr<RialtoConnector>  rialtoConnector;
+#endif //ENABLE_RIALTO_FEATURE
         };
 
         struct PluginData